refactor(search): decompose search-index.ts into focused modules (#226)
