+++ .devin/wiki.json
@@ -55,7 +55,7 @@
     },
     {
       "title": "Search and Indexing",
-      "purpose": "The SQLite FTS5 + sqlite-vec search system in search/. Covers: search-index.ts (the factory/closure that creates the SQLite database, FTS5 virtual table with body + metadata columns, vec0 virtual table for 384-dim embeddings, indexing pipeline that extracts tags/folder/type/leading-callout/frontmatter, and query functions for full-text search, tag queries, folder browsing, property queries, recent notes, backlinks, outgoing links, and orphan detection); the embedding pipeline — embedder.ts (lazy-loading ONNX pipeline for bge-small-en-v1.5 with singleton guard against concurrent model downloads), chunker.ts (heading-aware chunking with paragraph sub-splitting for oversized sections and title prefixing), and content-hash gating (SHA-256 per chunk) that skips re-embedding unchanged content on both incremental updates and full rebuilds; a non_md_files table that tracks non-markdown vault files (.canvas, .base, images, PDFs) so wikilinks to assets resolve as kind: 'asset' instead of counting as broken — outgoing links carry a kind discriminator ('note' vs 'asset') and broken-link counts exclude resolved assets; file-watcher.ts (chokidar watcher that keeps the FTS index, vector embeddings, and non_md_files table current on every file add/change/unlink, with per-path promise serialization for embedding and Windows polling mode support). The FTS index is rebuilt from scratch on startup; vector tables persist across restarts with content-hash gating skipping unchanged chunks. Embedding runs as a background Pass 3 so the server starts accepting FTS queries immediately. Hybrid search (hybridSearch) runs FTS5 keyword search then vector KNN sequentially, fuses results via Reciprocal Rank Fusion (computeRrfScores, k=60 with top-rank bonuses), and falls back to FTS-only silently when vectors are unavailable. vectorSearch embeds the query, runs sqlite-vec KNN, and deduplicates to the best chunk per note. The response includes search_mode ('hybrid' or 'fts') so clients know which ranking produced the scores. The vault_search tool description is conditional on EMBEDDING_ENABLED — hybrid-aware when true, FTS-only when false.",
+      "purpose": "The SQLite FTS5 + sqlite-vec search system in search/, decomposed across five modules. search-index.ts is the factory/closure (createSearchIndex) — schema, migrations, prepared statements, write operations (upsertNote, embedAndStoreChunks, removeNote, rebuildFromVault), non-md file tracking, and context wiring that binds query functions to the db instance. search-queries.ts holds all 15 query methods (fullTextSearch, hybridSearch, searchByTag, searchByFolder, recentNotes, listPropertyKeys, listPropertyValues, searchByProperty, getBacklinks, getOutgoingLinks, findOrphans, brokenLinkCount, modifiedOnDate, vaultStats, listAllTags) as named exports taking a SearchQueryContext. search-helpers.ts contains pure data transforms — row mappers (rowToMetadata, noteRowToSearchResult), filters (noteMatchesSearchFilters), type guards (isString, coerceToArray), and utilities (buildFtsMetadataText, mtimeToIso, buildSnippetFromChunkText, escapeLikeWildcards). fts-query.ts encapsulates FTS5 query sanitization (sanitizeFtsQuery). rrf.ts implements Reciprocal Rank Fusion scoring (computeRrfScores, k=60 with top-rank bonuses). The embedding pipeline — embedder.ts (lazy-loading ONNX pipeline for bge-small-en-v1.5 with singleton guard against concurrent model downloads), chunker.ts (heading-aware chunking with paragraph sub-splitting for oversized sections and title prefixing), and content-hash gating (SHA-256 per chunk) that skips re-embedding unchanged content on both incremental updates and full rebuilds. A non_md_files table tracks non-markdown vault files (.canvas, .base, images, PDFs) so wikilinks to assets resolve as kind: 'asset' instead of counting as broken. file-watcher.ts (chokidar watcher that keeps the FTS index, vector embeddings, and non_md_files table current on every file add/change/unlink, with per-path promise serialization for embedding and Windows polling mode support). Hybrid search runs FTS5 keyword search then vector KNN sequentially, fuses results via RRF, and falls back to FTS-only silently when vectors are unavailable. The vault_search tool description is conditional on EMBEDDING_ENABLED — hybrid-aware when true, FTS-only when false.",
       "parent": "Architecture"
     },
     {
